requirements.txt
South==0.8.4
django-debug-toolbar==1.0.1
selenium==2.40.0
